requirements.txt
discord.py>=2.0.0
openai==0.28.0
python-dotenv>=0.19.0
nest-asyncio>=1.5.1
wikipedia>=1.4.0
tiktoken>=0.5.1
beautifulsoup4>=4.9.3
requests>=2.25.1
google-api-python-client>=2.0.0
youtube-transcript-api>=0.6.0
pandas>=2.0.0 